Remove the local prettify and point to CDN
Put the CDN in correct file lol

refactored the square, box, and circle sizing to use classes like button-box-large and 'flattened' to top level style. Also, removed some button types which won't be in the 'lite' verion

Flatten the colors for the raised (so button-raised button-primary, now needs to be button-raised button-raised-primary). This is so that we can flatten the rules.

Lots more hacking to flatten styles. Removed a few more 'types' that didn't seem easy to flatten. Updated showcase